+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+
+### 3.0
+- Added the yaars related endpoint
+- Added database Tables and relationships
+- Added REST endpoints for all Tables
+- Implemented websocket endpoints for yaars related communication 
+
 ### 2.2.0
 - Added settings controller